ajuste en usuario

+++ src/app/services/redirect/redirect.service.ts
@@ -2,6 +2,8 @@ import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
 import Swal from 'sweetalert2';
 import { LoginService } from '../login/login.service';
+import { Profile } from '../../models/profile/profile';
+import { locale } from 'moment';
 
 @Injectable({
   providedIn: 'root'
@@ -11,7 +13,17 @@ export class RedirectService {
   constructor(private router: Router, private loginSevice: LoginService) { }
 
   login() {
-    localStorage.clear();
+    
+    if(localStorage.length > 0){
+      let userName = JSON.parse(localStorage.getItem('userName'));
+      localStorage.clear();
+      this.loginSevice.logOut(userName).subscribe((response: any) => {
+      },
+        error => {
+          console.log(JSON.stringify(error));
+      });
+    }
+
     this.router.navigate(['login']).then(() => {
       Swal.fire({
         icon: 'warning',
@@ -22,18 +34,12 @@ export class RedirectService {
         window.location.reload();
       });
     });
+
   }
 
 
   logout() {
 
-    this.loginSevice.logOut().subscribe((response: any) => {
-    },
-      error => {
-        console.log(JSON.stringify(error));
-    });
-
-    localStorage.clear();
     this.router.navigate(['login']).then(() => {
       Swal.fire({
         icon: 'success',
@@ -44,6 +50,15 @@ export class RedirectService {
         window.location.reload();
       });
     });
+
+    let userName = JSON.parse(localStorage.getItem('userName'));
+    localStorage.clear();
+    this.loginSevice.logOut(userName).subscribe((response: any) => {
+    },
+      error => {
+        console.log(JSON.stringify(error));
+    });
+
   }
 
   register() {
